Add files via upload
This is a html parser that I made in Python and it can take traverse the html file given top it and find the data in between a specific start tag and end with specific attributes given by the start tag. Once the program gets the information, it prints out "Vocab Word: " followed by the data found between the start and end tag in the html file.