Create Website.html
This is a small html file that I created that shows my expertise in html and java script. If this file is opened in a web browser, it will be a website with a link to this git hub and it will have a button that changes the top header when clicked on.