Create api.cpp
This project was created to mimic the office hours queue that we have at the University of Michigan for the Computer Science students. The data from the queue is read in as json data and information is then passed back to the online queue from the api in json format as well.